# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DraftAI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/settings/platform": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Read Platform Config",
        "operationId": "read_platform_config_api_v1_settings_platform_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformConfigPublic"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Update Platform Config",
        "operationId": "update_platform_config_api_v1_settings_platform_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformConfigPublic"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/ai/models": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Read Ai Models Catalog",
        "operationId": "read_ai_models_catalog_api_v1_settings_ai_models_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ollama_base_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 2048
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ollama Base Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIModelsCatalogPublic"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/ai": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Read Ai Config",
        "operationId": "read_ai_config_api_v1_settings_ai_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIAgentConfigPublic"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Update Ai Config",
        "operationId": "update_ai_config_api_v1_settings_ai_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIAgentConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIAgentConfigPublic"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/ai/providers": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Create Custom Ai Provider",
        "operationId": "create_custom_ai_provider_api_v1_settings_ai_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomAIProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomAIProviderPublic"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/ai/providers/{provider_id}": {
      "delete": {
        "tags": [
          "Settings"
        ],
        "summary": "Delete Custom Ai Provider",
        "operationId": "delete_custom_ai_provider_api_v1_settings_ai_providers__provider_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/ai/providers/{provider_slug}/refresh-models": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Refresh Provider Models",
        "operationId": "refresh_provider_models_api_v1_settings_ai_providers__provider_slug__refresh_models_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Slug"
            }
          },
          {
            "name": "ollama_base_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 2048
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ollama Base Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderModelsRefreshPublic"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIAgentConfigPublic": {
        "properties": {
          "provider": {
            "type": "string",
            "maxLength": 64,
            "title": "Provider",
            "default": "openai"
          },
          "model": {
            "type": "string",
            "maxLength": 128,
            "title": "Model",
            "default": "gpt-4o"
          },
          "cover_image_model": {
            "type": "string",
            "maxLength": 128,
            "title": "Cover Image Model",
            "default": "gemini-2.5-flash-image"
          },
          "temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 0.7
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8000
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "saved_api_keys": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Saved Api Keys"
          },
          "ollama_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ollama Base Url"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "AIAgentConfigPublic"
      },
      "AIAgentConfigUpdate": {
        "properties": {
          "provider": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "cover_image_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Model"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8000
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          },
          "ollama_base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Ollama Base Url"
          },
          "provider_api_keys": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Api Keys"
          },
          "openai_api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Openai Api Key"
          },
          "anthropic_api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Anthropic Api Key"
          },
          "gemini_api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Gemini Api Key"
          },
          "groq_api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Groq Api Key"
          },
          "openrouter_api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Openrouter Api Key"
          }
        },
        "type": "object",
        "title": "AIAgentConfigUpdate"
      },
...
      "AIModelsCatalogPublic": {
        "properties": {
          "providers": {
            "items": {
              "$ref": "#/components/schemas/AIProviderCatalogPublic"
            },
            "type": "array",
            "title": "Providers"
          },
          "cover_image_models": {
            "items": {
              "$ref": "#/components/schemas/CoverImageModelOption"
            },
            "type": "array",
            "title": "Cover Image Models"
          },
          "default_cover_image_model": {
            "type": "string",
            "title": "Default Cover Image Model"
          },
          "credential_providers": {
            "items": {
              "$ref": "#/components/schemas/AIProviderCredentialPublic"
            },
            "type": "array",
            "title": "Credential Providers"
          },
          "ollama": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OllamaCatalogStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "providers",
          "cover_image_models",
          "default_cover_image_model",
          "credential_providers"
        ],
        "title": "AIModelsCatalogPublic"
      },
...
      "CustomAIProviderCreate": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 128,
            "title": "Display Name"
          },
          "base_url": {
            "type": "string",
            "maxLength": 2048,
            "title": "Base Url"
          },
          "api_adapter": {
            "type": "string",
            "maxLength": 64,
            "title": "Api Adapter",
            "description": "openai_compatible or ollama",
            "default": "openai_compatible"
          },
          "auth_style": {
            "type": "string",
            "maxLength": 32,
            "title": "Auth Style",
            "default": "bearer"
          },
          "auth_header_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Header Name"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "display_name",
          "base_url"
        ],
        "title": "CustomAIProviderCreate"
      },
      "CustomAIProviderPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "api_adapter": {
            "type": "string",
            "title": "Api Adapter"
          },
          "auth_style": {
            "type": "string",
            "title": "Auth Style"
          },
          "requires_api_key": {
            "type": "boolean",
            "title": "Requires Api Key"
          },
          "has_api_key": {
            "type": "boolean",
            "title": "Has Api Key"
          },
          "models_source": {
            "type": "string",
            "title": "Models Source",
            "default": "live"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "display_name",
          "base_url",
          "api_adapter",
          "auth_style",
          "requires_api_key",
          "has_api_key"
        ],
        "title": "CustomAIProviderPublic"
      },
...
      "ErrorDetail": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "A specific error message suitable for UI display"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorDetail"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlatformConfigPublic": {
        "properties": {
          "devto_enabled": {
            "type": "boolean",
            "title": "Devto Enabled",
            "default": true
          },
          "linkedin_enabled": {
            "type": "boolean",
            "title": "Linkedin Enabled",
            "default": true
          },
          "devto_profile_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Devto Profile Url"
          },
          "linkedin_profile_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin Profile Url"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "has_devto_api_key": {
            "type": "boolean",
            "title": "Has Devto Api Key",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "PlatformConfigPublic"
      },
      "PlatformConfigUpdate": {
        "properties": {
          "devto_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Devto Enabled"
          },
          "linkedin_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin Enabled"
          },
          "devto_profile_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Devto Profile Url"
          },
          "linkedin_profile_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin Profile Url"
          },
          "devto_api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Devto Api Key"
          }
        },
        "type": "object",
        "title": "PlatformConfigUpdate"
      },
      "ProviderModelsRefreshPublic": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/AIModelOption"
            },
            "type": "array",
            "title": "Models"
          },
          "added_count": {
            "type": "integer",
            "title": "Added Count",
            "default": 0
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "default": 0
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "models"
        ],
        "title": "ProviderModelsRefreshPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}